dwim configuration settings
rather than passing an augeas set to configure the PHP ini, we try to be
smarter about the configuration settings. They can now be set in the
title as "key=value", or as key => value pairs.

n.b.: This first commit only fixes code & tests to comply with the
documentation, it does *not* yet test the second code path.